+++ src/test/java/com/zepben/evolve/services/network/testdata/SingleLineSwerNetwork.kt
@@ -10,7 +10,7 @@ package com.zepben.evolve.services.network.testdata
 
 import com.zepben.evolve.cim.iec61970.base.core.PhaseCode
 import com.zepben.evolve.services.network.NetworkService
-import com.zepben.evolve.services.network.testdata.TestDataCreators.*
+import com.zepben.evolve.services.network.testdata.*
 
 object SingleLineSwerNetwork {