+++ src/test/kotlin/com/zepben/evolve/streaming/get/testdata/FeederNetworkWithCurrent.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 Zeppelin Bend Pty Ltd
+ * Copyright 2021 Zeppelin Bend Pty Ltd
  *
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -16,10 +16,10 @@ import com.zepben.evolve.services.network.NetworkService
 import com.zepben.evolve.services.network.testdata.*
 import com.zepben.evolve.services.network.tracing.Tracing
 
-object FeederNetwork {
-    //             c1       c2
-    // source-fcb------fsp------tx
-    //
+object FeederNetworkWithCurrent {
+    //                c1       c2      c3       c4
+    //    source-fcb------fsp------tx------sw--------tx2
+    //                                   (open)
     fun create(): NetworkService {
         val networkService = NetworkService()
 
@@ -27,9 +27,14 @@ object FeederNetwork {
         val fcb = createSwitchForConnecting(networkService, "fcb", 2, PhaseCode.AB)
         val fsp = createNodeForConnecting(networkService, "fsp", 2, PhaseCode.AB)
         val tx = createPowerTransformerForConnecting(networkService, "tx", 2, PhaseCode.AB, 0, 0)
+        val tx2 = createPowerTransformerForConnecting(networkService, "tx2", 2, PhaseCode.AB, 0, 0)
+        val sw = createSwitchForConnecting(networkService, "sw", 2, PhaseCode.AB)
+        sw.setOpen(true)
 
         val c1 = createAcLineSegmentForConnecting(networkService, "c1", PhaseCode.AB)
         val c2 = createAcLineSegmentForConnecting(networkService, "c2", PhaseCode.AB)
+        val c3 = createAcLineSegmentForConnecting(networkService, "c3", PhaseCode.AB)
+        val c4 = createAcLineSegmentForConnecting(networkService, "c4", PhaseCode.AB)
 
         val substation = createSubstation(networkService, "f", "f", null)
         createFeeder(networkService, "f001", "f001", substation, fsp, fsp.getTerminal(2))
@@ -41,14 +46,22 @@ object FeederNetwork {
         addLocation(networkService, fcb, listOf(1.0, 1.0))
         addLocation(networkService, fsp, listOf(5.0, 1.0))
         addLocation(networkService, tx, listOf(10.0, 2.0))
+        addLocation(networkService, sw, listOf(15.0, 3.0))
+        addLocation(networkService, tx2, listOf(20.0, 4.0))
         addLocation(networkService, c1, listOf(1.0, 1.0, 5.0, 1.0))
         addLocation(networkService, c2, listOf(5.0, 1.0, 10.0, 2.0))
+        addLocation(networkService, c3, listOf(10.0, 1.0, 15.0, 3.0))
+        addLocation(networkService, c4, listOf(15.0, 1.0, 20.0, 4.0))
 
         networkService.connect(source.getTerminal(1)!!, fcb.getTerminal(1)!!)
         networkService.connect(fcb.getTerminal(2)!!, c1.getTerminal(1)!!)
         networkService.connect(c1.getTerminal(2)!!, fsp.getTerminal(1)!!)
         networkService.connect(fsp.getTerminal(2)!!, c2.getTerminal(1)!!)
         networkService.connect(c2.getTerminal(2)!!, tx.getTerminal(1)!!)
+        networkService.connect(tx.getTerminal(2)!!, c3.getTerminal(1)!!)
+        networkService.connect(c3.getTerminal(2)!!, sw.getTerminal(1)!!)
+        networkService.connect(sw.getTerminal(2)!!, c4.getTerminal(1)!!)
+        networkService.connect(c4.getTerminal(2)!!, tx2.getTerminal(1)!!)
 
         Tracing.setPhases().run(networkService)
         Tracing.assignEquipmentContainersToFeeders().run(networkService)