+++ changelog.md
@@ -4,8 +4,26 @@
  None.
 
 ##### New Features
-* None.
+* NetworkConsumerClient has 4 new functions:
+    - For fetching equipment for an EquipmentContainer
 
+          getEquipmentForContainer(service: NetworkService, mrid: String)         
+    - For fetching current equipment for a Feeder
+
+          getCurrentEquipmentForFeeder(service: NetworkService, mrid: String)         
+    - For fetching equipment for an OperationalRestriction
+
+          getEquipmentForRestriction(service: NetworkService, mrid: String)         
+    - For fetching terminals for a ConnectivityNode
+
+          getTerminalsForConnectivityNode(service: NetworkService, mrid: String)         
+
+* `NetworkConsumerClient.get_feeder()` now resolves all references, and thus you can expect to receive a Feeder with all equipment and their associations populated.
+
+* BaseService has two new functions which allow retrieving the UnresolvedReferences for an mRID by either `toMrid` or `from.mRID`: 
+  - getUnresolvedReferencesTo(mRID: String): Sequence<UnresolvedReference<*,*>>
+  - getUnresolvedReferencesFrom(mRID: String): Sequence<UnresolvedReference<*,*>>
+    
 ##### Enhancements
 * None.